+++ app/src/organisms/Labware/helpers/getAllDefs.ts
@@ -1,9 +1,9 @@
-import type { LabwareDefinition2 } from '../types'
+import type { LabwareDefinition2 } from '@opentrons/shared-data'
 
 // require all definitions in the labware/definitions/2 directory
 // require.context is webpack-specific method
 const definitionsContext = require.context(
-  '../../labware/definitions/2',
+  '@opentrons/shared-data/labware/definitions/2',
   true, // traverse subdirectories
   /\.json$/, // import filter
   'sync' // load every definition into one synchronous chunk
