+++ shared-data/js/index.ts
@@ -4,6 +4,5 @@ export * from './helpers'
 export * from './pipettes'
 export * from './types'
 export * from './labwareTools'
-export * from './labwareTools/definitions'
 export * from './modules'
 export * from '../protocol'
